pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "orchestra-llm-cache"
version = "0.2.1"
authors = [
  { name="Orchestra Team", email="team@orchestra.ai" },
]
description = "Reduce AI orchestration costs by 85% with semantic caching"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "numpy>=1.24.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.4",
]

[project.optional-dependencies]
langgraph = ["langgraph>=0.2.0"]
langchain = ["langchain>=0.1.0"]
full = ["langgraph>=0.2.0", "langchain>=0.1.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
